Move knmstate reporter to new package
