+++ test/e2e/utils.go
@@ -1,7 +1,6 @@
 package e2e
 
 import (
-	"bytes"
 	"context"
 	"fmt"
 	"github.com/nmstate/kubernetes-nmstate/build/_output/bin/go/src/encoding/json"
@@ -27,6 +26,7 @@ import (
 	dynclient "sigs.k8s.io/controller-runtime/pkg/client"
 
 	nmstatev1alpha1 "github.com/nmstate/kubernetes-nmstate/pkg/apis/nmstate/v1alpha1"
+	runner "github.com/nmstate/kubernetes-nmstate/test/e2e/runner"
 )
 
 const ReadTimeout = 180 * time.Second
@@ -207,48 +207,15 @@ func deletePolicy(name string) {
 
 }
 
-func run(command string, quiet bool, arguments ...string) (string, error) {
-	cmd := exec.Command(command, arguments...)
-	if !quiet {
-		GinkgoWriter.Write([]byte(command + " " + strings.Join(arguments, " ") + "\n"))
-	}
-	var stdout, stderr bytes.Buffer
-	cmd.Stderr = &stderr
-	cmd.Stdout = &stdout
-	err := cmd.Run()
-	if !quiet {
-		GinkgoWriter.Write([]byte(fmt.Sprintf("stdout: %.500s...\n, stderr %s\n", stdout.String(), stderr.String())))
-	}
-	return stdout.String(), err
-}
-
-func runAtNodeWithExtras(node string, quiet bool, command ...string) (string, error) {
-	ssh_command := []string{node, "--"}
-	ssh_command = append(ssh_command, command...)
-	output, err := run("./kubevirtci/cluster-up/ssh.sh", quiet, ssh_command...)
-	// Remove first two lines from output, ssh.sh add garbage there
-	outputLines := strings.Split(output, "\n")
-	output = strings.Join(outputLines[2:], "\n")
-	return output, err
-}
-
-func runQuiteAtNode(node string, command ...string) (string, error) {
-	return runAtNodeWithExtras(node, true, command...)
-}
-
-func runAtNode(node string, command ...string) (string, error) {
-	return runAtNodeWithExtras(node, false, command...)
-}
-
 func restartNode(node string) error {
 	By(fmt.Sprintf("Restarting node %s", node))
 	// Sync and reboot in background another way command can stuck
-	_, err := runAtNode(node, "/usr/bin/nohup /usr/bin/bash -c '/usr/bin/sync && sudo /usr/sbin/reboot -nf' > /dev/null 2>&1 &")
+	_, err := runner.RunAtNode(node, "/usr/bin/nohup /usr/bin/bash -c '/usr/bin/sync && sudo /usr/sbin/reboot -nf' > /dev/null 2>&1 &")
 	Expect(err).ToNot(HaveOccurred())
 	By(fmt.Sprintf("Waiting till node %s is rebooted", node))
 	// It will wait till uptime -p will return up that means that node was currently rebooted and is 0 min up
 	Eventually(func() string {
-		output, err := runAtNode(node, "uptime", "-p")
+		output, err := runner.RunAtNode(node, "uptime", "-p")
 		if err != nil {
 			return "not yet"
 		}
@@ -258,42 +225,11 @@ func restartNode(node string) error {
 	return nil
 }
 
-func kubectl(arguments ...string) (string, error) {
-	return run("./kubevirtci/cluster-up/kubectl.sh", false, arguments...)
-}
-
-func nmstatePods() ([]string, error) {
-	output, err := kubectl("get", "pod", "-n", framework.Global.Namespace, "--no-headers=true", "-o", "custom-columns=:metadata.name")
-	ExpectWithOffset(1, err).ToNot(HaveOccurred())
-	names := strings.Split(strings.TrimSpace(output), "\n")
-	return names, err
-}
-
-func runAtPods(arguments ...string) {
-	nmstatePods, err := nmstatePods()
-	ExpectWithOffset(1, err).ToNot(HaveOccurred())
-	for _, nmstatePod := range nmstatePods {
-		exec := []string{"exec", "-n", framework.Global.Namespace, nmstatePod, "--"}
-		execArguments := append(exec, arguments...)
-		_, err := kubectl(execArguments...)
-		ExpectWithOffset(1, err).ToNot(HaveOccurred())
-	}
-}
-
-func runAtNodes(command ...string) (outputs []string, errs []error) {
-	for _, node := range nodes {
-		output, err := runAtNode(node, command...)
-		outputs = append(outputs, output)
-		errs = append(errs, err)
-	}
-	return outputs, errs
-}
-
 func deleteBridgeAtNodes(bridgeName string, ports ...string) []error {
 	By(fmt.Sprintf("Delete bridge %s", bridgeName))
-	_, errs := runAtNodes("sudo", "ip", "link", "del", bridgeName)
+	_, errs := runner.RunAtNodes(nodes, "sudo", "ip", "link", "del", bridgeName)
 	for _, portName := range ports {
-		_, slaveErrors := runAtNodes("sudo", "nmcli", "con", "delete", bridgeName+"-"+portName)
+		_, slaveErrors := runner.RunAtNodes(nodes, "sudo", "nmcli", "con", "delete", bridgeName+"-"+portName)
 		errs = append(errs, slaveErrors...)
 	}
 	return errs
@@ -301,21 +237,21 @@ func deleteBridgeAtNodes(bridgeName string, ports ...string) []error {
 
 func createDummyAtNodes(dummyName string) []error {
 	By(fmt.Sprintf("Creating dummy %s", dummyName))
-	_, errs := runAtNodes("sudo", "nmcli", "con", "add", "type", "dummy", "con-name", dummyName, "ifname", dummyName, "ip4", "192.169.1.50/24")
-	_, upErrs := runAtNodes("sudo", "nmcli", "con", "up", dummyName)
+	_, errs := runner.RunAtNodes(nodes, "sudo", "nmcli", "con", "add", "type", "dummy", "con-name", dummyName, "ifname", dummyName, "ip4", "192.169.1.50/24")
+	_, upErrs := runner.RunAtNodes(nodes, "sudo", "nmcli", "con", "up", dummyName)
 	errs = append(errs, upErrs...)
 	return errs
 }
 
 func deleteConnectionAtNodes(name string) []error {
 	By(fmt.Sprintf("Delete connection %s", name))
-	_, errs := runAtNodes("sudo", "nmcli", "con", "delete", name)
+	_, errs := runner.RunAtNodes(nodes, "sudo", "nmcli", "con", "delete", name)
 	return errs
 }
 
 func deleteDeviceAtNode(node string, name string) error {
 	By(fmt.Sprintf("Delete device %s  at node %s", name, node))
-	_, err := runAtNode(node, "sudo", "nmcli", "device", "delete", name)
+	_, err := runner.RunAtNode(node, "sudo", "nmcli", "device", "delete", name)
 	return err
 }
 
@@ -389,7 +325,7 @@ func toUnstructured(y string) interface{} {
 }
 
 func bridgeVlansAtNode(node string) (string, error) {
-	return runAtNode(node, "sudo", "bridge", "-j", "vlan", "show")
+	return runner.RunAtNode(node, "sudo", "bridge", "-j", "vlan", "show")
 }
 func getVLANFlagsEventually(node string, connection string, vlan int) AsyncAssertion {
 	By(fmt.Sprintf("Getting vlan filtering flags for node %s connection %s and vlan %d", node, connection, vlan))
@@ -477,7 +413,7 @@ func vlansCardinality(node string, connection string) AsyncAssertion {
 
 func bridgeDescription(node string, bridgeName string) AsyncAssertion {
 	return Eventually(func() (string, error) {
-		return runAtNode(node, "sudo", "ip", "-d", "link", "show", "type", "bridge", bridgeName)
+		return runner.RunAtNode(node, "sudo", "ip", "-d", "link", "show", "type", "bridge", bridgeName)
 	}, ReadTimeout, ReadInterval)
 }
